fix: 管理画面プレビューを公開ページと構造一致させる
セクションコンポーネント・TOC・バッジ等を @ghost/shared に移動し、
web-public は re-export に変更。preview-content.tsx を shared コンポーネント
で書き換え、TOC（デスクトップ/モバイル）・スクロールスパイ・
CaseFileHeader・DetailSidebar を公開ページと同一構成にする。

- 15 コンポーネントを shared に移動（DRY 原則）
- web-public: 13 ファイルを re-export 化（-1024 行削減）
- web-admin: preview-content.tsx を shared コンポーネントで再構築
- 両アプリ tsc --noEmit 通過確認済み
