+++ packages/shared/package.json
@@ -8,6 +8,8 @@
     "firebase": ">=12.0.0",
     "react": ">=19.0.0",
     "react-dom": ">=19.0.0",
-    "next": ">=16.0.0"
+    "next": ">=16.0.0",
+    "react-markdown": ">=9.0.0",
+    "remark-gfm": ">=4.0.0"
   }
 }
